+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- Simplified Chinese translations!
+
 ## [1.10.1] - 2018-02-10
 ### Fixed
 - 12 hour time not displaying correctly in some locales.